Add tegra lib stubs and remove unused nvjpeg requirement
NVIDIA does not distribute the jetson multimedia libraries other than on
Jetson board images, so the only way to link against these libs is to
compile directly on a Jetson, making it difficult to create automated
builds or Docker images. This commit adds stubs of those libraries so
that nvmpi and ffmpeg can be built without them. The real libraries will
still be required at runtime.